% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_LifeRStyle.R
\name{anova_table}
\alias{anova_table}
\title{Extract an ANOVA table from a fitted LifeRstyle model}
\usage{
anova_table(fit_obj)
}
\arguments{
\item{fit_obj}{A fitted model object of class \code{"st403_fit"}.}
}
\value{
A data frame containing the ANOVA table.
}
\description{
Extract an ANOVA table from a fitted LifeRstyle model
}
\examples{
\dontrun{
fit_obj <- fit_lifeRstyle(HIS15_cleaned, fit_type = "anova")
anova_table(fit_obj)
}
}
