% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleaned_datasets}
\alias{cleaned_datasets}
\alias{HIS01_cleaned}
\alias{HIS09_cleaned}
\alias{HIS15_cleaned}
\alias{HSPAO11_cleaned}
\alias{alcohol_cleaned}
\alias{combined_cleaned}
\alias{health_cleaned}
\alias{smoking_cleaned}
\title{Cleaned LifeRstyle survey datasets}
\format{
Each object is a data frame.

An object of class \code{data.frame} with 468 rows and 5 columns.

An object of class \code{data.frame} with 936 rows and 5 columns.

An object of class \code{data.frame} with 1170 rows and 5 columns.

An object of class \code{data.frame} with 130 rows and 5 columns.

An object of class \code{data.frame} with 1890 rows and 5 columns.

An object of class \code{data.frame} with 4410 rows and 6 columns.

An object of class \code{data.frame} with 840 rows and 5 columns.

An object of class \code{data.frame} with 1680 rows and 5 columns.
}
\source{
Cleaned internally from raw survey data.
}
\usage{
HIS01_cleaned

HIS09_cleaned

HIS15_cleaned

HSPAO11_cleaned

alcohol_cleaned

combined_cleaned

health_cleaned

smoking_cleaned
}
\description{
A collection of cleaned survey datasets used in the LifeRstyle project.
}
\keyword{datasets}
