% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_LifeRStyle.R
\name{download_and_clean_cso}
\alias{download_and_clean_cso}
\title{Download and clean any table from CSO}
\usage{
download_and_clean_cso(
  table_id,
  dest_file = NULL,
  filter_sex = NULL,
  filter_age = NULL,
  filter_years = NULL
)
}
\arguments{
\item{table_id}{An array of numbers which corresponds to a specific data set on the CSO website.}

\item{dest_file}{The destination of the downloaded and cleaned data set.}

\item{filter_sex}{A parameter to filter the sex column of the data set.}

\item{filter_age}{A parameter to filter the age column of the data set.}

\item{filter_years}{A parameter to filter the years of which the data was.}
}
\value{
An object of class \code{"lifeRstyle"} which is a list of \code{\link[tibble]{tibble}}s which
contain the downloaded and cleaned data from the CSO website.
}
\description{
This function downloads and cleans data downloaded from the CSO website.
}
\examples{
\dontrun{
alcohol <- download_and_clean_cso("HIS15_cleaned")
}
}
