% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_LifeRStyle.R
\name{download_clean_combine_cso}
\alias{download_clean_combine_cso}
\title{Download, clean, and combine multiple CSO tables}
\usage{
download_clean_combine_cso(
  table_ids,
  filter_sex = NULL,
  filter_age = NULL,
  filter_years = NULL,
  combine = TRUE,
  save_dir = "data/clean"
)
}
\arguments{
\item{table_ids}{An array of numbers which corresponds to a specific data
set on the CSO website.This parameter can accept multiple \code{table_id}'s.}

\item{filter_sex}{A parameter to filter the sex column of the data set.}

\item{filter_age}{A parameter to filter the age column of the data set.}

\item{filter_years}{A parameter to filter the years of which the data was.}

\item{combine}{An operator set to \code{"TRUE"} by default which combines all specified \code{table_ids}.}

\item{save_dir}{Directory where cleaned CSV files will be saved}
}
\value{
An object of class \code{"lifeRstyle"} which is a list of \code{\link[tibble]{tibble}}s which
contain the downloaded,cleaned and combined data from the CSO website.
}
\description{
This function downloads,cleans and combines data sets downloaded from the CSO website.
}
\examples{
\dontrun{
# Named vector of tables
tables <- c(alcohol = "HIS15_cleaned", health = "HIS01", smoking = "HIS09")

# Download, clean, and combine
data_list <- download_clean_combine_cso(
tables,
filter_sex = NULL,   # set to c("Both sexes") if you want
filter_age = NULL,   # set to c("All ages") if needed
filter_years = NULL  # set to c("2019", "2020") if needed
)
# Access combined dataset
combined_data <- data_list$combined
dim(combined_data)
head(combined_data)
}
}
