% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_LifeRStyle.R
\name{plot_combined_data}
\alias{plot_combined_data}
\title{Create plots for combined CSO data}
\usage{
plot_combined_data(data)
}
\arguments{
\item{data}{A cleaned combined LifeRstyle dataset.}
}
\value{
A list containing three ggplot objects: histogram, boxplot, and scatter plot.
}
\description{
Generates a histogram, boxplot, and scatter plot for combined CSO data.
}
\examples{
plots <- plot_combined_data(combined_cleaned)
plots$histogram
}
