% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_LifeRStyle.R
\name{st403_fit-methods}
\alias{st403_fit-methods}
\alias{print.st403_fit}
\alias{summary.st403_fit}
\alias{plot.st403_fit}
\title{Methods for st403_fit objects}
\usage{
\method{print}{st403_fit}(x, ...)

\method{summary}{st403_fit}(object, ...)

\method{plot}{st403_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{st403_fit}.}

\item{...}{Further arguments passed to methods.}

\item{object}{An object of class \code{st403_fit}.}
}
\description{
Methods for st403_fit objects
}
